Fix view loading on startup
